+++ Sources/xcresultparser/CoberturaCoverageConverter.swift
@@ -47,7 +47,9 @@ public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
     override public func xmlString(quiet: Bool) throws -> String {
         let dtd = readDTD()
         dtd.name = "coverage"
-        dtd.systemID = "http://cobertura.sourceforge.net/xml/coverage-04.dtd"
+        // dtd.systemID = "http://cobertura.sourceforge.net/xml/coverage-04.dtd"
+        dtd.systemID =
+        "https://github.com/cobertura/cobertura/blob/master/cobertura/src/site/htdocs/xml/coverage-04.dtd"
 
         let rootElement = makeRootElement()
 
@@ -144,15 +146,11 @@ public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
     }
 
     private func readDTD() -> XMLDTD {
-        if let dtdUrl = URL(string: "http://cobertura.sourceforge.net/xml/coverage-04.dtd"),
-           let dtd = try? XMLDTD(contentsOf: dtdUrl) {
-            return dtd
+        do {
+            return try XMLDTD(data: Data(dtd04.utf8))
+        } catch {
+            fatalError("DTD could not be constructed. Error: \(error.localizedDescription)")
         }
-        if let dtdUrl = Bundle.module.url(forResource: "coverage-04", withExtension: "dtd"),
-           let dtd = try? XMLDTD(contentsOf: dtdUrl) {
-            return dtd
-        }
-        fatalError("DTD could not be constructed")
     }
 
     private func makeRootElement() -> XMLElement {
@@ -178,6 +176,77 @@ public class CoberturaCoverageConverter: CoverageConverter, XmlSerializable {
 
         return rootElement
     }
+    
+    // this ised to be fetched online from http://cobertura.sourceforge.net/xml/coverage-04.dtd
+    // that broke, when the URL changed to:
+    // https://github.com/cobertura/cobertura/blob/master/cobertura/src/site/htdocs/xml/coverage-04.dtd
+    // In case we couldn't download the data, we had a file as fallback. However that file could never be read
+    // because as command line tool this is not a bundle and thus there is no file to be found in the bundle
+    // IMO all that was overengineered for the followong 60 lines string...
+    // ...which will probably never ever change!
+    private var dtd04 = """
+<!-- Portions (C) International Organization for Standardization 1986:
+     Permission to copy in any form is granted for use with
+     conforming SGML systems and applications as defined in
+     ISO 8879, provided this notice is included in all copies.
+-->
+
+  <!ELEMENT coverage (sources?,packages)>
+  <!ATTLIST coverage line-rate        CDATA #REQUIRED>
+  <!ATTLIST coverage branch-rate      CDATA #REQUIRED>
+  <!ATTLIST coverage lines-covered    CDATA #REQUIRED>
+  <!ATTLIST coverage lines-valid      CDATA #REQUIRED>
+  <!ATTLIST coverage branches-covered CDATA #REQUIRED>
+  <!ATTLIST coverage branches-valid   CDATA #REQUIRED>
+  <!ATTLIST coverage complexity       CDATA #REQUIRED>
+  <!ATTLIST coverage version          CDATA #REQUIRED>
+  <!ATTLIST coverage timestamp        CDATA #REQUIRED>
+
+  <!ELEMENT sources (source*)>
+
+  <!ELEMENT source (#PCDATA)>
+
+  <!ELEMENT packages (package*)>
+
+  <!ELEMENT package (classes)>
+  <!ATTLIST package name        CDATA #REQUIRED>
+  <!ATTLIST package line-rate   CDATA #REQUIRED>
+  <!ATTLIST package branch-rate CDATA #REQUIRED>
+  <!ATTLIST package complexity  CDATA #REQUIRED>
+
+  <!ELEMENT classes (class*)>
+
+  <!ELEMENT class (methods,lines)>
+  <!ATTLIST class name        CDATA #REQUIRED>
+  <!ATTLIST class filename    CDATA #REQUIRED>
+  <!ATTLIST class line-rate   CDATA #REQUIRED>
+  <!ATTLIST class branch-rate CDATA #REQUIRED>
+  <!ATTLIST class complexity  CDATA #REQUIRED>
+
+  <!ELEMENT methods (method*)>
+
+  <!ELEMENT method (lines)>
+  <!ATTLIST method name        CDATA #REQUIRED>
+  <!ATTLIST method signature   CDATA #REQUIRED>
+  <!ATTLIST method line-rate   CDATA #REQUIRED>
+  <!ATTLIST method branch-rate CDATA #REQUIRED>
+  <!ATTLIST method complexity  CDATA #REQUIRED>
+
+  <!ELEMENT lines (line*)>
+
+  <!ELEMENT line (conditions*)>
+  <!ATTLIST line number CDATA #REQUIRED>
+  <!ATTLIST line hits   CDATA #REQUIRED>
+  <!ATTLIST line branch CDATA "false">
+  <!ATTLIST line condition-coverage CDATA "100%">
+
+  <!ELEMENT conditions (condition*)>
+
+  <!ELEMENT condition EMPTY>
+  <!ATTLIST condition number CDATA #REQUIRED>
+  <!ATTLIST condition type CDATA #REQUIRED>
+  <!ATTLIST condition coverage CDATA #REQUIRED>
+"""
 }
 
 private struct LineInfo {